float int from_range

+++ src/modm/math/algorithm/prescaler.hpp
@@ -136,12 +136,12 @@ public:
 			   T desired_frequency,
 			   uint32_t begin, uint32_t end)
 	{
-		const double desired = double(input_frequency) / desired_frequency;
+		const float desired = static_cast<float>(input_frequency / desired_frequency);
 		const uint32_t prescaler_floor = std::max(uint32_t(std::floor(desired)), begin);
 		const uint32_t prescaler_ceiling = std::min(uint32_t(std::ceil(desired)), end);
 		const T baud_lower = input_frequency / prescaler_ceiling;
 		const T baud_upper = input_frequency / prescaler_floor;
-		const double baud_middle = (baud_upper + baud_lower) / 2.0;
+		const float baud_middle = static_cast<float>((baud_upper + baud_lower) / 2.0);
 		if (desired_frequency <= baud_middle)
 			return {input_frequency, desired_frequency, prescaler_ceiling - begin, prescaler_ceiling};
 		return {input_frequency, desired_frequency, prescaler_floor - begin, prescaler_floor};
